Renamed EndToEndTest to EndToEndTestAddAssignment. Created two test cases. Where all add assignment form is filled correctly, and one where it is filled incorrectly. Both test case have passed.
